+++ lib/emerge/runtime/viewport/renderer.ex
@@ -18,8 +18,6 @@ defmodule Emerge.Runtime.Viewport.Renderer do
   @callback set_input_target(term(), pid() | nil) :: :ok
   @callback set_log_target(term(), pid() | nil) :: :ok
   @callback set_input_mask(term(), non_neg_integer()) :: :ok
-  @callback submit_video_frame(term(), atom(), VideoInterop.Frame.t()) ::
-              :ok | {:error, term()}
 
   @callback upload_tree(term(), Emerge.Engine.Element.t()) ::
               {Emerge.Engine.diff_state(), Emerge.Engine.Element.t()}
@@ -30,6 +28,5 @@ defmodule Emerge.Runtime.Viewport.Renderer do
   @callback patch_tree_runtime(term(), Emerge.Engine.diff_state(), Emerge.Engine.Element.t()) ::
               {Emerge.Engine.diff_state(), nil}
 
-  @optional_callbacks patch_tree_runtime: 3,
-                      submit_video_frame: 3
+  @optional_callbacks patch_tree_runtime: 3
 end